address_space: Deduplicate {A32,A64}AddressSpace
