+++ src/dynarmic/CMakeLists.txt
@@ -373,6 +373,8 @@ elseif(ARCHITECTURE STREQUAL "arm64")
         backend/arm64/a32_jitstate.h
         backend/arm64/abi.cpp
         backend/arm64/abi.h
+        backend/arm64/address_space.cpp
+        backend/arm64/address_space.h
         backend/arm64/devirtualize.h
         backend/arm64/emit_arm64.cpp
         backend/arm64/emit_arm64.h
