+++ src/dynarmic/backend/arm64/a32_address_space.cpp
@@ -14,6 +14,7 @@
 #include "dynarmic/common/fp/fpcr.h"
 #include "dynarmic/frontend/A32/a32_location_descriptor.h"
 #include "dynarmic/frontend/A32/translate/a32_translate.h"
+#include "dynarmic/interface/A32/config.h"
 #include "dynarmic/interface/exclusive_monitor.h"
 #include "dynarmic/ir/opt/passes.h"
 
@@ -97,9 +98,8 @@ static void* EmitExclusiveWriteCallTrampoline(oaknut::CodeGenerator& code, const
 }
 
 A32AddressSpace::A32AddressSpace(const A32::UserConfig& conf)
-        : conf(conf)
-        , mem(conf.code_cache_size)
-        , code(mem.ptr()) {
+        : AddressSpace(conf.code_cache_size)
+        , conf(conf) {
     EmitPrelude();
 }
 
@@ -119,33 +119,6 @@ IR::Block A32AddressSpace::GenerateIR(IR::LocationDescriptor descriptor) const {
     Optimization::VerificationPass(ir_block);
 
     return ir_block;
-}
-
-CodePtr A32AddressSpace::Get(IR::LocationDescriptor descriptor) {
-    if (const auto iter = block_entries.find(descriptor.Value()); iter != block_entries.end()) {
-        return iter->second;
-    }
-    return nullptr;
-}
-
-CodePtr A32AddressSpace::GetOrEmit(IR::LocationDescriptor descriptor) {
-    if (CodePtr block_entry = Get(descriptor)) {
-        return block_entry;
-    }
-
-    IR::Block ir_block = GenerateIR(descriptor);
-    const EmittedBlockInfo block_info = Emit(std::move(ir_block));
-
-    block_infos.insert_or_assign(descriptor.Value(), block_info);
-    block_entries.insert_or_assign(descriptor.Value(), block_info.entry_point);
-    return block_info.entry_point;
-}
-
-void A32AddressSpace::ClearCache() {
-    block_entries.clear();
-    block_infos.clear();
-    block_references.clear();
-    code.set_ptr(prelude_info.end_of_prelude);
 }
 
 void A32AddressSpace::EmitPrelude() {
@@ -291,18 +264,8 @@ void A32AddressSpace::EmitPrelude() {
     mem.protect();
 }
 
-size_t A32AddressSpace::GetRemainingSize() {
-    return conf.code_cache_size - (code.ptr<CodePtr>() - reinterpret_cast<CodePtr>(mem.ptr()));
-}
-
-EmittedBlockInfo A32AddressSpace::Emit(IR::Block block) {
-    if (GetRemainingSize() < 1024 * 1024) {
-        ClearCache();
-    }
-
-    mem.unprotect();
-
-    const EmitConfig emit_conf{
+EmitConfig A32AddressSpace::GetEmitConfig() {
+    return EmitConfig{
         .optimizations = conf.unsafe_optimizations ? conf.optimizations : conf.optimizations & all_safe_optimizations,
 
         .hook_isb = conf.hook_isb,
@@ -328,132 +291,6 @@ EmittedBlockInfo A32AddressSpace::Emit(IR::Block block) {
 
         .coprocessors = conf.coprocessors,
     };
-    EmittedBlockInfo block_info = EmitArm64(code, std::move(block), emit_conf);
-
-    Link(block.Location(), block_info);
-
-    mem.invalidate(reinterpret_cast<u32*>(block_info.entry_point), block_info.size);
-
-    RelinkForDescriptor(block.Location());
-
-    mem.protect();
-
-    return block_info;
-}
-
-static void LinkBlockLinks(const CodePtr entry_point, const CodePtr target_ptr, const std::vector<BlockRelocation>& block_relocations_list) {
-    using namespace oaknut;
-    using namespace oaknut::util;
-
-    for (auto [ptr_offset] : block_relocations_list) {
-        CodeGenerator c{reinterpret_cast<u32*>(entry_point + ptr_offset)};
-
-        if (target_ptr) {
-            c.B((void*)target_ptr);
-        } else {
-            c.NOP();
-        }
-    }
-}
-
-void A32AddressSpace::Link(IR::LocationDescriptor block_descriptor, EmittedBlockInfo& block_info) {
-    using namespace oaknut;
-    using namespace oaknut::util;
-
-    for (auto [ptr_offset, target] : block_info.relocations) {
-        CodeGenerator c{reinterpret_cast<u32*>(block_info.entry_point + ptr_offset)};
-
-        switch (target) {
-        case LinkTarget::ReturnToDispatcher:
-            c.B(prelude_info.return_to_dispatcher);
-            break;
-        case LinkTarget::ReturnFromRunCode:
-            c.B(prelude_info.return_from_run_code);
-            break;
-        case LinkTarget::ReadMemory8:
-            c.BL(prelude_info.read_memory_8);
-            break;
-        case LinkTarget::ReadMemory16:
-            c.BL(prelude_info.read_memory_16);
-            break;
-        case LinkTarget::ReadMemory32:
-            c.BL(prelude_info.read_memory_32);
-            break;
-        case LinkTarget::ReadMemory64:
-            c.BL(prelude_info.read_memory_64);
-            break;
-        case LinkTarget::ExclusiveReadMemory8:
-            c.BL(prelude_info.exclusive_read_memory_8);
-            break;
-        case LinkTarget::ExclusiveReadMemory16:
-            c.BL(prelude_info.exclusive_read_memory_16);
-            break;
-        case LinkTarget::ExclusiveReadMemory32:
-            c.BL(prelude_info.exclusive_read_memory_32);
-            break;
-        case LinkTarget::ExclusiveReadMemory64:
-            c.BL(prelude_info.exclusive_read_memory_64);
-            break;
-        case LinkTarget::WriteMemory8:
-            c.BL(prelude_info.write_memory_8);
-            break;
-        case LinkTarget::WriteMemory16:
-            c.BL(prelude_info.write_memory_16);
-            break;
-        case LinkTarget::WriteMemory32:
-            c.BL(prelude_info.write_memory_32);
-            break;
-        case LinkTarget::WriteMemory64:
-            c.BL(prelude_info.write_memory_64);
-            break;
-        case LinkTarget::ExclusiveWriteMemory8:
-            c.BL(prelude_info.exclusive_write_memory_8);
-            break;
-        case LinkTarget::ExclusiveWriteMemory16:
-            c.BL(prelude_info.exclusive_write_memory_16);
-            break;
-        case LinkTarget::ExclusiveWriteMemory32:
-            c.BL(prelude_info.exclusive_write_memory_32);
-            break;
-        case LinkTarget::ExclusiveWriteMemory64:
-            c.BL(prelude_info.exclusive_write_memory_64);
-            break;
-        case LinkTarget::CallSVC:
-            c.BL(prelude_info.call_svc);
-            break;
-        case LinkTarget::ExceptionRaised:
-            c.BL(prelude_info.exception_raised);
-            break;
-        case LinkTarget::InstructionSynchronizationBarrierRaised:
-            c.BL(prelude_info.isb_raised);
-            break;
-        case LinkTarget::AddTicks:
-            c.BL(prelude_info.add_ticks);
-            break;
-        case LinkTarget::GetTicksRemaining:
-            c.BL(prelude_info.get_ticks_remaining);
-            break;
-        default:
-            ASSERT_FALSE("Invalid relocation target");
-        }
-    }
-
-    for (auto [target_descriptor, list] : block_info.block_relocations) {
-        block_references[target_descriptor.Value()].emplace(block_descriptor.Value());
-        LinkBlockLinks(block_info.entry_point, Get(target_descriptor), list);
-    }
-}
-
-void A32AddressSpace::RelinkForDescriptor(IR::LocationDescriptor target_descriptor) {
-    for (auto block_descriptor : block_references[target_descriptor.Value()]) {
-        if (auto iter = block_infos.find(block_descriptor); iter != block_infos.end()) {
-            const EmittedBlockInfo& block_info = iter->second;
-
-            LinkBlockLinks(block_info.entry_point, Get(target_descriptor), block_infos[block_descriptor].block_relocations[target_descriptor]);
-
-            mem.invalidate(reinterpret_cast<u32*>(block_info.entry_point), block_info.size);
-        }
-    }
 }
 
 }  // namespace Dynarmic::Backend::Arm64
